[netstack] Signal connectedness for accepted sockets
Fix a race where we can skip signalling connectedness on the zircon
socket corresponding to an accepted endpoint in an error state.

On `Accept` while creating a new endpoint, if the gVisor state has a
hard error, we asynchronously invoke endpoint close which races with
checking for `closing` before signalling connectedness. This can cause
the connectedness signalling to be skipped causing the client to
incorrectly interpret errors on the accepted socket as unconnected.

Fix the race by always signalling connectedness for accepted sockets
before handling any hard errors on the endpoint.

Bug: 67142
Test: fx test -od netstack_bsdsocket_c_api_test -- \
--gtest_filter=LocalhostTest.AcceptAfterReset --gtest_repeat=100000

Change-Id: Ie1b34e69840dde8821078673c8a401984f5fec6a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502051
Commit-Queue: Mithun Iyer <iyerm@google.com>
Reviewed-by: Tamir Duberstein <tamird@google.com>