[kernel][lockup] Fix DAP powered backtraces
Fix the backtraces during fatal condition violations detected by the
lockup detector that we produce using register dumps harvested by the
DAP.  The problem was that the shadow call stack register (x18) is
always pointing to the _next_ place to stash a return address, not at
the last address stashed.  In addition, the link register (x30) is
what holds the immediate return address of the current frame at all
times.

Because of this, frame 0 of the backtrace was correct, as were frames
2-N, but frame 1 was showing some stale data on the shadow stack,
while LR held the real value.

Fix this issue by appropriately dumping the LR as frame one, and
attempting to back up in the stack before producing frames 2 and
onward.

Bug: 70724
Change-Id: I895829469d61286e02ab9b2885a6cd6f6575e626
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502711
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>