[media][tools] Extend rates, improve confidence
This CL builds on the recent extention of downsampling ratios,
by changing the profiled "extreme" ratios from 8:192 and 192:8
to 1:192 and 192:1. The 192:192 ratio is added.

To improve confidence intervals for profiler measurements, the
default number of iterations is increased from 500 to 1000.

Execution times are slightly increased:
* For the perftest subset, from 6 sec to 8 sec
* For the full developer set, from 180 sec to 245 sec

Change-Id: I523225a308ae4109e8bdab637bb367d050dcc6e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501244
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>