[build][sdk] Add coverage_sdk_source_sets build argument.
This build argument can be set to true when the "profile" variant
is selected in order to restrict profiling to the sdk_source_set()
instances reachable from //sdk:sdk (as well as anything listed
optionally through 'profile_source_files' and
'profile_source_files_list_files').

This will speed up coverage analysis for SDK source libraries
tremendously on infra, since the CTS tests, and the rest of the
platform can be built uninstrumented and run normally.

IMPORTANT: This CL introduces an non-trivial behaviour change in
our build system: everytime the list of sources used for all
sdk_source_set() in our source tree changes, the
//sdk/sdk_source_set.profile_list will need to be updated.

This list should change rarely over time, and a mechanism is
introduced to check when this is needed and print a human
friendly error message that explains how to fix the situation
easily.

For more details, read the comments in //sdk/config/profile/sdk.gni
which explains the implementation and how this is supposed to be
used!

For testing:

  fx set core.x64 --with //sdk --args 'coverage_sdk_source_sets = true'
  fx build

Alternatively, remove one line from //sdk/sdk_source_set.profile_list
and re-run 'fx build' to see the error message that instruct you to
copy the generated file to the right location. Follow the instructions
then run 'fx build' again (which will be fast).

Bug: 68618

Change-Id: Ibc4aa3780b9a1b531528a9c3e7edb57f5b541637
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497938
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: David Turner <digit@google.com>