[scenic] Make ViewTreeUpdater a singleton
Move the ViewTreeUpdater out of gfx::Session and into gfx::Engine
as a singleton. All Sessions (and AnnotationManager) now add their
updates to this ViewTreeUpdater, and all updates are applied together
in a single call at the end of GfxSystem::UpdateSessions().

This solves the problem of ViewTreeUpdates not being applied if the
owning session doesn't have scheduled updates.

Added tests for both bugs.

Fixed: 56345, 59413
Test: fx test gfx_unittests input_unittests

Change-Id: I4151dab7e57b25c7ce3d961200b4472ca744167e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500942
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>
Reviewed-by: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Yilong Li <liyl@google.com>