[stress_tests] Store last operation in FVM stress tests
FVM stress test operations are not atomic. It is possible
for an operation to complete partially before a crash occurs.

If the volume actor fails to perform an operation because
of a volume disconnect, store the operation so that it can
be performed again when the volume is reconnected.

Fixed: 70180
Test: fx test fvm-crash
      fx test fvm-nocrash
Change-Id: I8efc8e6362ddee4f2f705bf1792b27af717dc124
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484100
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>