[block_client] Convert RemoteBlockDevice to trait
Turns `struct RemoteBlockDevice` into `trait BlockClient` and
`struct RemoteBlockClient`.

This is pre-work to create an in-process fake implementation of
BlockClient which can be used for in-process tests. (Currently clients
have to stand up a remote ramdisk to test against this class, which is
heavyweight.)

Leaves RemoteBlockDevice aliased to facilitate a soft migration.

Change-Id: I2204942de6730bcfadef4275e879fc5b5304d7ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502918
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>