[zircon][scripts] Only rebuild the JSON file when generating docs.
`update-docs-from-fidl` pulls information from a generated JSON file.
If the output JSON file does not exist or is out of date, the script
will perform an `fx build`.

Because the input VDSO files are deep in the build dependencies, each
change to the input file typically requires recompilin a large amount of
the system. This makes iterating on documentation painfully slow.

Update the script to only build the JSON file, and not the entire
system.

Change-Id: Ifdeeaa511b883f73b94b2ee7df0f7d8d2faec76d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501439
Reviewed-by: Scott Graham <scottmg@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: David Greenaway <dgreenaway@google.com>