[media/vnext] update fidl libraries for media vnext prototyping
This CL updates the fidl libraries to //src/media/vnext/fidl. These
definitions are a 'work in progress', but are checked in so that
members of the media team can write code against them to refine and
expand on them. The intention is that they will eventually move into
//sdk/fidl and become part of the sdk.

Definitions in fuchsia.media2 and fuchsia.mediastreams are based on
design documents currently under review. Definitions in fuchsia.audio
and fuchsia.mem2 have not yet been subject to review and are
speculative. fuchsia.mem2.BufferProvider, in particular, is a 'toy'
alternative to sysmem and it missing many features.

Test: no tests - fidl definitions only
Change-Id: I0a593a1ca400547fd95b6ee7cdcd65546ce8db25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503581
Reviewed-by: Tom Bergan <tombergan@google.com>
API-Review: Dale Sather <dalesat@google.com>
Commit-Queue: Dale Sather <dalesat@google.com>