[lib/arch][x86] Introduce X86MsrBase
This is a convenience class to inherit from in defining an MSR register
type. It saves on boilerplate (especially the eyesore of a static cast
of the associated X86Msr enum value back to uint32_t) and mirrors the
base types we have introduced for defining system and CPUID registers.

Change-Id: Idc00072de563dc971d1a2681746c69a84857ec0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501662
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>