[kernel][phys][x86] Set up GDT in physboot.
When physboot is entered, we are in 64-bit mode but don't know the
current state of x86-specific state such as the GDT.

This CL adds a new architecture-specific `ArchSetUp` function. On x86,
this sets up x86 Global Descriptor Table and Task Segment data
structures and registers.

As part of this change, we fill out lib/arch's 64-bit GDT descriptor
and helper fuctions for loading system registers.

Change-Id: I35864b869dd954faef3dcc848294983524c3fa10
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/495409
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>