[ffx] plumb ssh port from target to host pipe
An optional field is added to the target that enables us to override
the ssh port on a per target basis. This port is plumbed through to
the host-pipe connection setup, but is not yet plumbed upward to
target add, which comes next.

Bug: 66457
Change-Id: Iecc736fec9ce0dc3238ae129eb9793e9649d36e4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/487857
Fuchsia-Auto-Submit: James Tucker <raggi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jordon Wing <jwing@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>