[fxfs] Add BlockDevice
BlockDevice is an instance of Device backed by a real block client (i.e.
a remote_block_device::BlockClient).

This also introduces VmoBufferSource, which is a single VMO that is
mapped and attached to the remote block device which all transfer
buffers are allocated out of.

Change-Id: I7a757f8941e35dc337c0d16e327cdfca9547aede
Tested: Unit.
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502739
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>