[flutter-embedder-test] Moves the injected services into a function.
Removes the use of a fixed-size array in injected services
initialization.

Needing to keep a manual count of the number of elements in the array is
error prone.  And, the use of the service list seems simple enough not
to really require the array optimization.

So, I propose to remove the size constraint, and replace with
std::vector.

Change-Id: Ic683cb990c1034b5a398043e3e7b06eeed6d42e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503780
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Sanjay Chouksey <sanjayc@google.com>
Reviewed-by: David Worsham <dworsham@google.com>