[build/go] Use `go mod vendor`
Care is taken to match versions to what is currently pinned in the
integration repository. Updating those dependencies to latest will
happen in a subsequent change.

One exception to the above: cloud.google.com/go was updated from
3f2988d48e25 to ad4f9324cdd7. The latter revision removed honnef.co/go
from go.mod, which enables us to avoid vendoring that repository, which
we do not use.

Updated bootstrap.sh and fx-set to build in a temporary directory where
go modules are properly initialized. This also prevents `fx go` from
downloading sources by setting GOPROXY=off.

Note that vendoring takes place in third_party/golibs because the
behavior of `go mod vendor` is to nuke everything that isn't managed by
it, which interferes with the source layout of the fuchsia project. This
is tracked in https://github.com/golang/go/issues/44646.

Bug: 35565
Change-Id: If228b778ac747b9a9ebc1719ba31a8b67823e2ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503479
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Tamir Duberstein <tamird@google.com>