[wlan][brcmfmac] Expand the FIDL ReasonCode enum to cover all codes
The ReasonCode enum no longer omits reserved codes but instead maps
them into the nearest reserved code bucket. For example, if an AP
returns the reason code 754, then that code will now be mapped to
RESERVED_129_TO_65535.

The impact of this change is to remove the mapping of reserved codes
to UNSPECIFIED_REASON to avoid ambiguity in metrics. In particular,
other applications appears to receive frames with the reserved reason
code 0 very often, and its unclear how many times Fuchsia sees that
code as well.

Fixed: 72507
Test: fx test ieee80211-codes-tests
Change-Id: Ib90059696f9d11dfea9dbc31ffb5d9055f492078