[volume_image]: Add support for decompress image.
To simplify the tool, all images are decompressed. A compressed image is
decompressed before being dealt with. The function takes a Writer, which
means we can choose to decompress into a new file(temporary or not) or
in memory.

Test: storage-volume-image-tests
Change-Id: Ibda1c46b2006bd6336c63a290a91423c238964b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502867
Fuchsia-Auto-Submit: Gianfranco Valentino <gevalentino@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>