[scripts][fuzzing] Improve handling of non-directory corpus targets
In several places, we had implicitly assumed that GN corpus targets
always directly correspond to a directory in the source tree containing
corpus elements. This assumption does not hold true for corpus targets
of the form `//a:b`, where `b`≠`a`, so we did not handle these cases
correctly.

This CL modifies the way `fx fuzz` maps GN labels to package namespace
paths to match the approach taken in `fuzzer_test_component`. It also
ensures `fx fuzz update` gives an explicit error when run for fuzzers
with a non-directory corpus target, as automatic buildfile generation is
not possible in this case.

There remains a potential for name collisions between targets of the
form `//a:b` and `//a/b:b`, but this is not addressed in this CL.

Incidentally, the copyright year for generated corpus buildfiles has
been freed from being stuck in 2020. May this bode well for us all.

Bug: 71634
Change-Id: If9c081912a291fba90a126af63810833a1b1380a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/499878
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Aaron Green <aarongreen@google.com>
Commit-Queue: Mark Dittmer <markdittmer@google.com>