[vfs] Release the VMO observer.
Previously the new VFS pager (currently unused) did not "unwatch" for
pager events, leading to leaks. Unregister from pager events and test
that this happens.

Allow derived classes to control whether and when this unwatching
happens. This will allow blobfs to keep watching even when there are no
clones, which is requires to be able to use vmo_read for the FIDL read
path (which triggers the new pager to itself).

Change-Id: Iefc97895f9a4443e5da572d43539ec888b0d62be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/503300
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>