[wlan][policy] fix mixed security recording
This changes the security type returned from network selection to
reflect the security type of the saved network config, not of the actual
scan type.
The scan security type is used to match loaded saved networks with scan
results, and the actual scan security type is not used after network
selection, but the config security type is needed to record whether the
connection is successful. Past connect failures are used to choose when
there are multiple options for connecting.

Bug: 70964
Test: modified unit tests to reflect the security type that network
      selection should return:
	fx test wlancfg-tests
Change-Id: Ifd55fffbd946dc030659a0b5f6b333569c1bf987
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/490937
Reviewed-by: Kevin Sakuma <sakuma@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>
Commit-Queue: Naomi Mccracken <nmccracken@google.com>