[kernel][vm] Change relation between PageSource and PagerProxy
PagerProxy inherits directly from PageProvider. PageSource contains a
PageProvider instance, which is the PagerProxy sub-class for
pager-backed VMOs. This keeps the pager dispatcher related code on the
PagerProxy side separated out from the VMO handling code on the
PageSource side. To combine the functionality of the two, we use
composition instead of inheritance.

Change-Id: I54c2b164d20242fd01049767996b09c828aa07d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498481
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>