[fdt] update fx fdt due to cipd breakage
The CIPD upload script we rely on changed which broke our ability to
upload FDT using the same structure we had been using. To fix that we
had to change the layout of our package. This fixes the `fx fdt` script
to handle the new layout.

Change-Id: I1f7156f2d58c97212f0ee0e0ea1e6592f70b660d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504800
Commit-Queue: Andy Weiss <dragonbear@google.com>
Reviewed-by: Jayson Adams <shrike@google.com>
Reviewed-by: Alex Bogart <abbogart@google.com>