[fdevtools] update fdevtools due to cipd breakage
The CIPD upload script we rely on changed which broke our ability to
upload FDT using the same structure we had been using. To fix that we
had to change the layout of our package. This fixes the `fx fdt` script
to handle the new layout.

Change-Id: I732ff9c5d7b88db6300b6dd95793e02cb70c240f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504667
Reviewed-by: Jayson Adams <shrike@google.com>
Reviewed-by: Omer Lev-ran <omerlevran@google.com>
Commit-Queue: Andy Weiss <dragonbear@google.com>