[blobfs] Wire up blob creation to new pager.
Hooks up the blob creation path to the new pager. This avoids creating a
PageWatcher for the blob in the new path and all of the functionality is
now in the Blob object.

This involved creating a callback that handles the differences between
VMO creation between the pager and using that in the blob loader. A nice
side effect is that the pager dependency from the load is no longer
necessary.

To support removing the PageWatcher completely, some additional ifdefed
variables were added to the Blob object. This is unfortunate but makes
this more clear about what's currently happening.

The new paging system still does not net work properly and is off by
default so there should be no behavior change.

Change-Id: I7ff178f1ef42593cdbbcc5435c586666bea66dae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504238
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>