[camera] Disable free space probe on stream buffer allocation
The free space probe check before allocating the stream buffers was
causing failure to allocate when switching streams. There may be a
bug in the promise chain that waits for free space causing multiple
overlapping requests to stay active. It's not clear where that bug is
though so we disable for now to unblock testing of camera features.

No allocation failures were observed with this patch when switching
streams.

Test: See linked bugs
Bug: b/182385264, b/182905702
Change-Id: I660a36174fef95188f2d5accc38984d880d9938e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504816
Commit-Queue: Austin Foxley <afoxley@google.com>
Reviewed-by: Ernest Hua <ernesthua@google.com>