[build] Suppress lsan for vboot_reference
https://fxrev.dev/462880 created this allow-list of targets with LSAN
suppressed, but missed the target //tools/vboot_reference:cgpt, which
suppresses LSAN errors:
https://fuchsia.googlesource.com/fuchsia/+/master/tools/vboot_reference/BUILD.gn#144.

That allow-list change broke the cgpt target because it was no longer
allowed to suppress LSAN errors.

It seems that that target is not built in CQ, which is why the change
was able to land. However, the target *is* built on the host-prebuilts
builders that build tools and upload them to CIPD, so those builders
have been quietly failing for 3 months:
https://ci.chromium.org/p/fuchsia/builders/global.ci/host_prebuilts-x64-mac/b8852299601903060720

Bug: 72680
Change-Id: I919cdbac4f5cef04ca63c175c0685628f8b401cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504841
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Oliver Newman <olivernewman@google.com>