[tefmocheck] copy output files from tests to checks
When a string_in_log_check fails for a particular test, this will make
it much easier to find the root cause, since the LUCI and Sponge UIs
will show the per-test log which caused the check to fail. And this will
hopefully  help us understand the attached bug, which is currently
a mystery to me.

Bug: 72721
Change-Id: I83491c789c0cb1bdc22c17478d963751ab49825e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504939
Commit-Queue: Gary Miguel <garymm@google.com>
Fuchsia-Auto-Submit: Gary Miguel <garymm@google.com>
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>