[async] Always unregister pager watchers.
Previously if the watched VMO was destroyed before unregistering the
watcher for it, the async loop would fail and leave the watcher object
registered in the linked list of live objects. The caller generally
destroys the watcher object at this point, which results in a
use-after-free at message-loop teardown.

This change always unregisters the watcher object from the linked list,
even if VMO unregistration fails.

Change-Id: Ibc47731000e8c5cb32ee88b5cd3d7fa9bb960907
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504663
Reviewed-by: Corey Tabaka <eieio@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>