[blobfs] Implement read-only health checks
Adds a health check to the BlobfsVerifier service which verifies up to
1MiB of data (arbitrarily chosen to give reasonable confidence in health
checks without being too slow) for blobs in the cache.

We only pick blobs in the cache (i.e. those which have been opened since
the system booted) so that we are certain that we do not have false
negatives by verifying blobs that are not actively used. For instance,
if some ephemeral package was corrupt, we would not want it to fail the
health checks after an OTA.

This might mean that in practice few blobs are actually verified (since
we do the verification early in the boot), but false negatives are
significantly worse than false positives here, and at least we would
know that the important early blobs such as the binaries/libraries of
the update stack are valid.

Also did a bit of refactoring in blob.cc (noticed that LoadAndVerifyBlob
was only used for the null blob case).

Bug: 70920
Tested: Unit tests.
Change-Id: I356c18eb13cb9fd98abb6586835bd27d4fddacda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/502098
Commit-Queue: Zach Kirschenbaum <zkbaum@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>