[debugger] Support 0-based line table files.
DWARF defines the 0 file name index in the line table to implicitly
refer to the file name of the compilation unit. Before DWARF v5, Clang
did not emit this and we treated it as an invalid entry.

With DWARF v5 support, Clang started to emit this (even though it was
always valid). We now allow a 0 line table entry to support these
symbols.

Bug: 26814
Change-Id: Iaa22190e3d7ace04d07662876e99c2ab29496483
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504844
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>