[fxfs] Fix skip-list locking
Prior to this CL, the skip list unsafely used a RWLock. This CL
addresses that.  With this implementation, writers cannot block readers.
Writers are sequenced i.e. there can only be one writer at a time and
writers are guaranteed to eventually make progress (so long as readers
do).  The approach taken is similar to lock-free skip-list
implementations.

This CL also changes the iterator API so that seek now returns an
iterator rather than creating an iterator and then having to call seek.
This simplifies iterator implementations which no longer need to keep
track of the special state that exists prior to calling seek.

The merge API is still a little odd, but we can revisit that later.

MULTIPLY: fxfs-tests: 100
Change-Id: I0112fae4a2ded601e7ff4d2411877456f59914c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504007
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>