[ui-input][fidl] Replace confusing "stream" terminology.
Event streams are potentially infinite in length, but we sometimes must
refer to a specific event sequence of user engagement. To make that act
of reference more clear, we conceptually break up a stream into a
sequence of interactions, so that the API is able to refer to a specific
interaction. Each term follows a state machine laid out in
fuchsia.ui.pointer.EventPhase.

TEST=CQ

Bug: 69175
Change-Id: I1d3ea1666ba3e35c66313f7a7d885eef0ca677a3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/484552
Commit-Queue: Jaeheon Yi <jaeheon@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>
API-Review: Alice Neels <neelsa@google.com>