[bind] Add true and false statements
The `false` statement replaces the `abort` statement, and the `true`
statement is new. Both of these must be the only statement in a block
(either nested in an if/else, or the only statement in the bind rules.)

The `true` statement is semantically a no-op because a bind rule will
match if it reaches the end of all of its conditions. It is
syntactically useful for writing a bind rule that always matches, or in
rare cases where you might otherwise want to write an empty block in an
if/else statement.

To avoid giving the impression that bind rules are imperative programs,
rather than declarative conditions, these keywords must be the only
statement present in their block. If any statement followed a `true`
statement it might be easy to incorrectly assume that the statement
would be ignored.

Change-Id: Ia3f93c943a9c4a6a95f377a886640ab063f847ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/494600
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Alex Legg <alexlegg@google.com>