[fidlgen_cpp] Replace unit types with namespace enum in `Kind`
In order to identify FIDL declarations, we employed marker types and
took advantage of default initialization. This trick has created
confusion and prevented defining an interface like the following:

    type Decl interface {
        Kind() _SomeKind_  // What to put in "_SomeKind_"?
                           // It is polymorphic.
    }

It turns out changing this unit type pattern to use namespaced enums
is just as safe and readable, while allowing us to properly define
the Decl interface.

As part of this change, we store IR value themselves into Decl (and
implement the Decl interface on the IR structs), as opposed to storing a
pointer to the IR values. This has the nice side-effect of avoiding
unnecessary boxing (see
https://golang.org/doc/effective_go#pointers_vs_values). In particular,
the Decls should not be mutated by the templates.

Change-Id: I449af7306ea85d5a089af030ba6415263eb95678
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504555
Commit-Queue: Yifei Teng <yifeit@google.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>