[stress_tests] Make CM stress tests hub-based and multi-threaded
The component manager stress tests now use the hub, rather than
maintaining an in-memory representation of the component tree.

Each actor connects to the hub separately, traverses to a random point
in the hierarchy and creates/deletes a random component using the
exposed fuchsia.sys2.Realm protocol.

Errors encountered during traversal, creation and deletion are
silently ignored. The assumption is that the operation could not
complete because of the actions of another actor.

This leads to interesting scenarios where, for example, a component
being bound to in one thread is also being deleted in another.

There are now 10 identical actors that do such operations across the
topology, each on a different thread.

Test: fx test component-manager-stress-tests-alt
Change-Id: If10f7517ff698189d8d8d3c7cf90d6d568a49a96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/504218
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>