[Bind] Add bytecode data to zircon_driver_note
Add the bytecode version and count to
zircon_driver_note so that the DriverManager can
use the data to evaluate the new bytecode.

This CL also fixes an error in the
ZIRCON_DRIVER_BEGIN_PRIV_V2 macro where the byte
count is not populated correctly.

Bug: 69609
Change-Id: Idd1029ae681a97f2b54460926a4448437f99cb45
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505418
Reviewed-by: Alex Legg <alexlegg@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Sarah Chan <spqchan@google.com>