[wlan] Include BYTE in SSID and PSK length constant variable names
SSID is limited to at most 32 bytes and PSK must be 32 bytes. It
is worth including BYTE in the variable names MAX_SSID_BYTE_LEN and
WPA_PSK_BYTE_LEN to make clear that the limit should be applied to
bytes and not characters. In particular, since SSID can sometimes be
interpreted as a UTF-8 string, it's possible for the SSID length limit
to be misinterpreted.

Test: Pure refactor

Change-Id: Ia19a76e73b8c6f49175aa7e394d38256df3571ea
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/501942
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Charles Celerier <chcl@google.com>
Reviewed-by: Marc Khouri <mnck@google.com>