[tefmocheck] simplify splitLogByTest().
In investigating the attached bug, I discovered that the per-test logs
are sometimes incorrect. However I haven't been able to reproduce the
issue locally, so I'm just trying to simplify the code in the hopes of
eliminating bugs. Even if this doesn't eliminate the issue, it seems
like an improvement in readability and performance.

Bug: 72721
Change-Id: Icc6f362a79de9576706ae7812bbca4df2d9f148d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505567
Commit-Queue: Gary Miguel <garymm@google.com>
Fuchsia-Auto-Submit: Gary Miguel <garymm@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>