[harvester] Add FIDL support
When harvester is run with the --fidl argument, it will now publish the
new fuchsia.systemmonitor.Harvester protocol. We've also added a new
component manifest to run with this argument, and configured it to
auto-start in sysmgr.

The protocol only has one method, ConnectGrpc, that takes a socket and
establishes a GRPC connection over which the legacy GRPC protocol can be
used.

Fixed: 69454
Change-Id: I8f80f18454ba60b3480c2b37ce9752e7912c8b69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500939
Reviewed-by: James Tucker <raggi@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Andy Weiss <dragonbear@google.com>
API-Review: James Tucker <raggi@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>