[netemul] Use `RealmOptions.name` for realm moniker
The optional `RealmOptions.name` is combined with two static
monotonically-increasing counters, for the sandbox instance and the
realm instance, to create a unique moniker for the root of the generated
realm.

This CL also introduces the `ManagedRealm.GetMoniker` method to enable
clients to retrieve the component moniker of the root of the runtime-
generated topology that contains the `ManagedRealm`'s constituent
components. This is necessary in order to test that the moniker is
correctly set.

Fixed: 72169
Test: fx test -d netemul-v2-tests --test-filter *realm_name*
Change-Id: I3eafd57c690b959797e3a8a017abaef1a4ec62f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505439
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
API-Review: Bruno Dal Bo <brunodalbo@google.com>