[ffx] Choose shortest tool path when getting a host tool
The SDK provides, typically, three paths for each tool we ship. Two of
them are arch-specific, and one of them is the "default" arch. We
previously just took whichever we happened to find first, but with this
change we'll always take the shortest, which we expect to be the default
one since usually the difference is that the path doesn't contain an
arch specifier.

The word "legacy" appears a lot around this "default" tool, so I suspect
this may break again soon. Looking in to a permanent fix.

Change-Id: I7e2e8d2a2312ba2b1ee27cd96fdb9ff83f69ac9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505842
Reviewed-by: Jordon Wing <jwing@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>