[time lib] Add a new Transform struct.
This is similar to the ClockTransform in ClockDetails in that it stores
the linear mapping from monotonic time to a synthetic time, but this
struct also stores the error bound and change in error bound over time.

This will be useful in implementing the timekeeper frequency correction
algorithm.

Bug: 56868
Change-Id: Id97065386d83ab2ac03df54dadb1a4ce0f673631
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505566
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Jody Sankey <jsankey@google.com>