[fxfs] Add reference types for Buffer
Adds BufferRef and MutableBufferRef, which are arbitrarily sliceable
reference types that are used for transacting with a Device
(read/write).

Reference types are necessary (instead of, e.g. &Buffer and &mut Buffer)
because we need to support slicing buffers in the
ObjectStore::read/write code. A slice of a buffer must necessarily yield
a reference of some kind, and adding a reference type was necessary to
make the lifetimes work out (since something needs to own the reference
and have the lifetimes appropriately scoped).

Buffer references must still be aligned by the time they are used for
transacting with the block device. Assertions are added to ensure that
we catch this at runtime.

Change-Id: I114d70640c7b7fa0584a0fc7840a450f79a736d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506164
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>