[scenic] Allow RootPresenter to be relaunched.
The main bug was that DisplaySwapchain::Flip() was being passed the
display id instead of the layer id.  Consequently, when a new
DisplaySwapchain was created with layer id == 2, it continued to
pass display id == 1 to Flip(), resulting in a black screen.

A secondary issue is that the layer was being destroyed while still
associated to a display.  The display controller logged this as an
error with no other apparent ill effects, but who knows what might
happen on other hardware?

Fixed: 73050
Change-Id: I961a5bb63937703fdd17c7a371ef5112f7243a07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506786
Reviewed-by: Yilong Li <liyl@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>