[ffx] Backend symbolization of proactive logs.
In this implementation we'll pass any log with a message that looks like
a context message to the `symbolizer` binary. We pass a known string
on a new line symbolizer to detect when a given input maps to zero or to
multiple lines of output.

Fixed: 71770
Change-Id: I058074e9f7a87265dcf26c5cda37c31c88ced6cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/498351
Commit-Queue: Jordon Wing <jwing@google.com>
Reviewed-by: Matthew Boetger <boetger@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>