[fit] Allow lib/fit in the kernel
Update the README text to account for the fit-promise split.
Also remove an obsolete public_configs setting that is not needed.

This includes a single migration of an fbl::MakeAutoCall use in the
kernel to fit::defer as demonstration fit in the kernel works now.

Fixed: 57350
Bug: 57355
Change-Id: Ia0f9b84b24e6c9f33d85324fed5c3c6b8cd117a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507082
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>