[triage] Make select support multiple sources in one.
Migrations of triage selectors to a new location is currently difficult.
It requires creating a parallel set of rules that check both locations a
metric may appear. This CL extends "select" clauses to support a vector
of locations, the first of which is used for the actual value.

Fixed: 62002
Change-Id: I4e39c9c584bda221de741518cc85bfb330b1bfda
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/440114
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Chris Phoenix <cphoenix@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>
Testability-Review: Adam Perry <adamperry@google.com>