[blobfs] Make fsck continue after errors
Fsck used to early exit when it detected corruption in the backup
superblock, which is not very useful since it hides the other issues
with the image. Change this to let fsck proceed through all checks
before returning whether the image is valid or not.

Tested: unit.
Change-Id: Ibfca83a2615bdf3521ec594eef7f7227c2dd8cbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506798
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>