[fidlgen][cpp] Improved name handling
Collapse type and decl handling into a single set of types.

Rename `DeclVariant` to `Name` and `DeclName` to `NameVariants`. It
turns out that in practice it's most common to name types and
declarations that are unique to a particular bindings variant, so the
API now centers on that.

Names themselves are represented in a tree of `namePart`s rather than an
array of strings. It's often very useful to be able to access a class'
name for itself - what it would call its constructor or destructor - and
due to how template specializations are named we need to introduce a
little more structure.

The main test for this change is that it doesn't cause a change in the
goldens.

Bug: 72798

Change-Id: I601457f95d6fe97718400081abfa241ff04d3455
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/506922
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>