[bt][gap] Restore bonds before starting scan
When the bonds are not restored before we start a passive scan (via
applying the configuration), restoring bonds can fail if we detect a
peer we are bonded to before the bonds are restored.

Restore the bonds before applying the config, ensuring this can't
happen.

Bug: b/182420131
Change-Id: Iea4cf40732538647e83c6e7f40411423e163a75a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507200
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>