[volume_image]: Blobfs Adapter.
Provides an adapter that generates the necessary metadata for processing
a blobfs image into a FvmDescriptor.

Added integration test for the image creation from input images.
Currently only blobfs.

This uncovered an OOB bug in the writing image code, added a regression
test for that as well.

Also uncovered a bug when writing a fvm image, where the mapping.source
offset was never applied when reading. This was not caught by the
existing tests, because the fake data never changed source to a non zero
value. No regression test added, but updated tests to trigger the bug,
and applied the fix.

Test: storage-volume-image-tests,
storage-volume-image-adapter-integration-tests

Change-Id: I9985b2030885caf18504c2a019f033637c21faf6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/505560
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Chris Suter <csuter@google.com>